Connection state management implemented
Created an atomic_int connect_state that keeps track of not yet
connected, connected, disconnected, and errors. Used to implement good
error behavior when the server is down and you start() a Redox instance,
or when the server goes down in the middle of running commands. Now,
nothing should hang, but should return errors (or throw exceptions) when
the server goes down.

Also added hooks for a user connect/disconnect callback in the
constructor, which is helpful for client programs.

Added an example with three Redox clients in one thread.