Binary value support for last word in command
Implemented limited but useful binary data support without breaking the
API. If the last character of the command is a ", look for the first ",
and everything in between the quotes treat as binary data. Takes care of
setting a key with a binary value. Not useful if there need to be
multiple binary entries in one command, or for binary keys. Also need to
be careful if the last character of the value actually needs to be a
quote, in which case we need to quote the value.