No-wait mode
User can now enable no-wait mode, which chooses whether we use the
EVRUN_NOWAIT flag in ev_run. The default is off, so that we don't
use 100% CPU. Note added in tutorial to enable when performance
is critical. Added to the speed test examples.

Bump to 0.2.1. Remove patch number from HISTORY entry - that's what the
git log is for. Make note on minor release.