Fix condition variable usage in subscriber
The lock used with the condition variable should always be the same one
guarding changes to the condition itself, in these cases, the lock
guarding changes to the subscribe and psubscibe data structures.

This commit also limits the number of locks held by the subscribeBase
function at any one time.